Rename JDBC integration tests
