+++ core/src/integration-test/java/com/scalar/db/storage/jdbc/JdbcDatabaseMultiplePartitionKeyIntegrationTest.java
@@ -3,11 +3,11 @@ package com.scalar.db.storage.jdbc;
 import com.scalar.db.config.DatabaseConfig;
 import com.scalar.db.io.DataType;
 import com.scalar.db.io.Value;
-import com.scalar.db.storage.StorageSinglePartitionKeyIntegrationTestBase;
+import com.scalar.db.storage.StorageMultiplePartitionKeyIntegrationTestBase;
 import java.util.Random;
 
-public class JdbcSinglePartitionKeyIntegrationTest
-    extends StorageSinglePartitionKeyIntegrationTestBase {
+public class JdbcDatabaseMultiplePartitionKeyIntegrationTest
+    extends StorageMultiplePartitionKeyIntegrationTestBase {
   private static RdbEngine rdbEngine;
 
   @Override
@@ -15,6 +15,14 @@ public class JdbcSinglePartitionKeyIntegrationTest
     JdbcConfig jdbcConfig = JdbcEnv.getJdbcConfig();
     rdbEngine = JdbcUtils.getRdbEngine(jdbcConfig.getContactPoints().get(0));
     return jdbcConfig;
+  }
+
+  @Override
+  protected int getThreadNum() {
+    if (rdbEngine == RdbEngine.ORACLE) {
+      return 1;
+    }
+    return super.getThreadNum();
   }
 
   @Override
