+++ core/src/integration-test/java/com/scalar/db/storage/jdbc/JdbcDatabaseSingleClusteringKeyScanIntegrationTest.java
@@ -3,11 +3,11 @@ package com.scalar.db.storage.jdbc;
 import com.scalar.db.config.DatabaseConfig;
 import com.scalar.db.io.DataType;
 import com.scalar.db.io.Value;
-import com.scalar.db.storage.StorageMultipleClusteringKeyScanIntegrationTestBase;
+import com.scalar.db.storage.StorageSingleClusteringKeyScanIntegrationTestBase;
 import java.util.Random;
 
-public class JdbcMultipleClusteringKeyScanIntegrationTest
-    extends StorageMultipleClusteringKeyScanIntegrationTestBase {
+public class JdbcDatabaseSingleClusteringKeyScanIntegrationTest
+    extends StorageSingleClusteringKeyScanIntegrationTestBase {
 
   private static RdbEngine rdbEngine;
 
@@ -16,14 +16,6 @@ public class JdbcMultipleClusteringKeyScanIntegrationTest
     JdbcConfig jdbcConfig = JdbcEnv.getJdbcConfig();
     rdbEngine = JdbcUtils.getRdbEngine(jdbcConfig.getContactPoints().get(0));
     return jdbcConfig;
-  }
-
-  @Override
-  protected int getThreadNum() {
-    if (rdbEngine == RdbEngine.ORACLE) {
-      return 1;
-    }
-    return super.getThreadNum();
   }
 
   @Override
